Serialize hebdo problem

+++ src/solver/optimisation/opt_optimisation_lineaire.cpp
@@ -146,7 +146,7 @@ bool runWeeklyOptimization(const OptimizationOptions& options,
                 copy(ncons, ProblemeHebdo->ProblemeAResoudre->SecondMembre, week_ptr->SecondMembre);
                 copy(ncons, ProblemeHebdo->ProblemeAResoudre->Sens, week_ptr->Sens);
 
-                lps->_hebdo[{year, week}] = week_ptr;
+                lps->_hebdo[{static_cast<unsigned int>(year), static_cast<unsigned int>(week)}] = week_ptr;
             }
         }
         /* Fin */
